requirements.txt
jinja2==2.11.3
jsonschema==3.2.0
pyyaml==5.4.1
